# src/swagger.yaml
openapi: 3.0.0
info:
  title: AI Sales API
  description: Complete API documentation for the AI Sales system.
  version: 1.0.0
security:
  - BearerAuth: []
paths:
...
  /api/v1/roles:
    post:
      tags: [Roles]
      summary: Create a new role
      description: Create a new role in the system.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreate'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Role with this name already exists

    get:
      tags: [Roles]
      summary: Get all roles
      description: Retrieve all roles in the system.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
  /api/v1/roles/{role_id}:
    get:
      tags: [Roles]
      summary: Get role by ID
      description: Retrieve a role by its unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          description: Role not found

    put:
      tags: [Roles]
      summary: Update role
      description: Update role details by its unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          description: Role not found

    delete:
      tags: [Roles]
      summary: Delete role
      description: Delete a role by its unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role deleted successfully
        '404':
          description: Role not found

  # Role Permissions Routes
...
  /api/v1/ai-personas:
    get:
      tags: [AI Personas]
      summary: Get all AI Personas
      description: Retrieve all AI Personas.
      responses:
        '200':
          description: AI Personas retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIPersonaResponse'
    post:
      tags: [AI Personas]
      summary: Create a new AI Persona
      description: Create a new AI Persona.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIPersonaCreate'
      responses:
        '201':
          description: AI Persona created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIPersonaResponse'
  /api/v1/ai-personas/{persona_id}:
    get:
      tags: [AI Personas]
      summary: Get AI Persona by ID
      description: Retrieve an AI Persona by its ID.
      parameters:
        - name: persona_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AI Persona retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIPersonaResponse'
        '404':
          description: AI Persona not found
    put:
      tags: [AI Personas]
      summary: Update AI Persona
      description: Update an AI Persona by its ID.
      parameters:
        - name: persona_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIPersonaCreate'
      responses:
        '200':
          description: AI Persona updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIPersonaResponse'
        '404':
          description: AI Persona not found
  /api/v1/ai-personas/industry/{industry}:
    get:
      tags: [AI Personas]
      summary: Get AI Personas by industry
      description: Retrieve all AI Personas for a specific industry.
      parameters:
        - name: industry
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: AI Personas retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIPersonaResponse'

  # Interaction Modes Routes
...
  /api/v1/ai-roles/:
    post:
      tags: [AI Roles]
      summary: Create a new AI role
      description: Create a new AI role in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRoleCreate'
      responses:
        '201':
          description: AI role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRoleResponse'
    get:
      tags: [AI Roles]
      summary: Get all AI roles
      description: Retrieve all AI roles in the system.
      responses:
        '200':
          description: AI roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIRoleResponse'
  /api/v1/ai-roles/{ai_role_id}:
    get:
      tags: [AI Roles]
      summary: Get AI role by ID
      description: Retrieve an AI role by its unique ID.
      parameters:
        - name: ai_role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AI role retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRoleResponse'
        '404':
          description: AI role not found
    put:
      tags: [AI Roles]
      summary: Update AI role
      description: Update an AI role by its unique ID.
      parameters:
        - name: ai_role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRoleUpdate'
      responses:
        '200':
          description: AI role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRoleResponse'
        '404':
          description: AI role not found
    delete:
      tags: [AI Roles]
      summary: Delete AI role
      description: Delete an AI role by its unique ID.
      parameters:
        - name: ai_role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: AI role deleted successfully
        '404':
          description: AI role not found

  # Industries Routes
  /api/v1/industries/:
    post:
      tags: [Industries]
      summary: Create a new industry
      description: Create a new industry in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndustryCreate'
      responses:
        '201':
          description: Industry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryResponse'
    get:
      tags: [Industries]
      summary: Get all industries
      description: Retrieve all industries in the system.
      responses:
        '200':
          description: Industries retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IndustryResponse'
  /api/v1/industries/{industry_id}:
    get:
      tags: [Industries]
      summary: Get industry by ID
      description: Retrieve an industry by its unique ID.
      parameters:
        - name: industry_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Industry retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryResponse'
        '404':
          description: Industry not found
    put:
      tags: [Industries]
      summary: Update industry
      description: Update an industry by its unique ID.
      parameters:
        - name: industry_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndustryUpdate'
      responses:
        '200':
          description: Industry updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryResponse'
        '404':
          description: Industry not found
    delete:
      tags: [Industries]
      summary: Delete industry
      description: Delete an industry by its unique ID.
      parameters:
        - name: industry_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Industry deleted successfully
        '404':
          description: Industry not found

  # Manufacturing Models Routes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoleBase:
      type: object
      properties:
        name:
          type: string
          enum:
            - sales_person
            - manager
            - admin
            - super_admin
        description:
          type: string
          nullable: true
    RoleCreate:
      allOf:
        - $ref: '#/components/schemas/RoleBase'
    RoleUpdate:
      type: object
      properties:
        name:
          type: string
          enum:
            - sales_person
            - manager
            - admin
            - super_admin
          nullable: true
        description:
          type: string
          nullable: true
    RoleResponse:
      allOf:
        - $ref: '#/components/schemas/RoleBase'
        - type: object
          properties:
            role_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            status_active:
              type: boolean

    # Role Permission Schemas
...
    AIPersonaCreate:
      type: object
      required:
        - name
        - industry_id
        - ai_role_id
        - experience_level
        - plant_size_impact_id
        - manufacturing_model_id
      properties:
        name:
          type: string
        industry_id:
          type: string
        ai_role_id:
          type: string
        experience_level:
          type: string
          enum: [junior, mid, senior]
          description: 'junior, mid, or senior'
        geography:
          type: string
        plant_size_impact_id:
          type: string
        manufacturing_model_id:
          type: string
        behavioral_traits:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              intensity:
                type: integer
              description:
                type: string
    AIPersonaResponse:
      type: object
      properties:
        persona_id:
          type: string
          format: uuid
        name:
          type: string
        industry_id:
          type: string
        ai_role_id:
          type: string
        experience_level:
          type: string
        geography:
          type: string
        plant_size_impact_id:
          type: string
        manufacturing_model_id:
          type: string
        behavioral_traits:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              intensity:
                type: integer
              description:
                type: string
        status_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AIRoleBase:
      type: object
      properties:
        name:
          type: string
    AIRoleCreate:
      allOf:
        - $ref: '#/components/schemas/AIRoleBase'
    AIRoleUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
    AIRoleResponse:
      allOf:
        - $ref: '#/components/schemas/AIRoleBase'
        - type: object
          properties:
            ai_role_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    IndustryBase:
      type: object
      properties:
        name:
          type: string
    IndustryCreate:
      allOf:
        - $ref: '#/components/schemas/IndustryBase'
    IndustryUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
    IndustryResponse:
      allOf:
        - $ref: '#/components/schemas/IndustryBase'
        - type: object
          properties:
            industry_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...